Dash Attack

+++ Assets/ALLPLAYTESTSTUFF/Scripts/CameraControl.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class CameraControl : MonoBehaviour {
     public GameObject player;
-    public int offset = -4;
+    public float offset = -4;
 	public Material fadeToBlack;
 	public float opacity = 0f;
 	Color temporary;
